middleware::LogRequests: Extract Sentry setup into dedicated middleware struct
